src/main/java/organice/model/person/Person.java
package organice.model.person;

import static organice.commons.util.CollectionUtil.requireAllNonNull;

import java.util.Objects;

/**
 * Represents a Person in ORGANice.
 * Guarantees: details are present and not null, field values are validated, immutable.
 */
public class Person {

    // Identity fields
    protected final Type type;
    protected final Nric nric;
    protected final Name name;
    protected final Phone phone;

    /**
     * Every field must be present and not null.
     */
    public Person(Type type, Nric nric, Name name, Phone phone) {
        requireAllNonNull(type, nric, name, phone);
        this.type = type;
        this.nric = nric;
        this.name = name;
        this.phone = phone;
    }

    public Type getType() {
        return type;
    }

    public Nric getNric() {
        return nric;
    }

    public Name getName() {
        return name;
    }

    public Phone getPhone() {
        return phone;
    }

    /**
     * Returns true if both persons of the same nric.
     * This defines a weaker notion of equality between two persons.
     */
    public boolean isSamePerson(Person otherPerson) {
        if (otherPerson == this) {
            return true;
        }

        return otherPerson != null
                && otherPerson.getNric().equals(getNric());
    }

    /**
     * Returns true if both persons have the same identity and data fields.
     * This defines a stronger notion of equality between two persons.
     */
    @Override
    public boolean equals(Object other) {
        if (other == this) {
            return true;
        }

        if (!(other instanceof Person)) {
            return false;
        }

        Person otherPerson = (Person) other;
        return otherPerson.getNric().equals(getNric())
                && otherPerson.getName().equals(getName())
                && otherPerson.getPhone().equals(getPhone())
                && otherPerson.getType().equals(getType());
    }

    @Override
    public int hashCode() {
        // use this method for custom fields hashing instead of implementing your own
        return Objects.hash(type, nric, name, phone);
    }

    @Override
    public String toString() {
        final StringBuilder builder = new StringBuilder();
        builder.append(getName())
                .append(" Person Type: ")
                .append(getType())
                .append(" Nric: ")
                .append(getNric())
                .append(" Phone: ")
                .append(getPhone());
        return builder.toString();
    }

}
